requirements.txt
aiofiles==23.1.0
annotated-types==0.5.0
anyio==3.6.2
attrs==22.2.0
bcrypt==4.0.1
bleach==6.0.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-log==0.4.0
colorama==0.4.6
contourpy==1.2.0
cryptography==41.0.7
cvxopt==1.3.2
cycler==0.12.1
deprecation==2.1.0
dnspython==2.3.0
docutils==0.20.1
dotty-dict==1.3.1
ecdsa==0.18.0
email-validator==2.1.0.post1
fastapi==0.109.0
fonttools==4.47.0
gitdb==4.0.10
GitPython==3.1.41
gotrue==2.1.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.18.0
httptools==0.6.1
httpx==0.25.0
idna==3.4
importlib-metadata==6.7.0
inflection==0.5.1
iniconfig==2.0.0
invoke==1.7.3
itsdangerous==2.1.2
jaraco.classes==3.2.3
jeepney==0.8.0
Jinja2==3.1.3
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
keyring==24.2.0
kiwisolver==1.4.5
MarkupSafe==2.1.2
matplotlib==3.8.2
more-itertools==9.1.0
numpy==1.24.2
openapi==1.1.0
orjson==3.9.12
packaging==23.0
pandas==1.5.3
passlib==1.7.4
patsy==0.5.5
Pillow==10.1.0
pkginfo==1.9.6
plotly==5.18.0
pluggy==1.0.0
postgrest==0.13.2
psycopg2-binary==2.9.6
pyasn1==0.5.0
pycparser==2.21
pydantic==2.5.3
pydantic-extra-types==2.4.1
pydantic-settings==2.1.0
pydantic_core==2.14.6
Pygments==2.15.1
pyparsing==3.1.1
pytest==7.2.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-gitlab==3.15.0
python-jose==3.3.0
python-multipart==0.0.6
python-semantic-release==7.33.2
pytz==2022.7.1
PyYAML==6.0.1
readme-renderer==40.0
realtime==1.0.0
referencing==0.32.0
requests==2.28.2
requests-toolbelt==1.0.0
rfc3986==1.5.0
rpds-py==0.15.2
rsa==4.9
scipy==1.11.4
seaborn==0.13.0
SecretStorage==3.3.3
semver==2.13.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
SQLAlchemy==2.0.19
starlette==0.35.1
statsmodels==0.14.1
storage3==0.7.0
StrEnum==0.4.15
supabase==2.3.3
supafunc==0.3.3
tenacity==8.2.3
tomlkit==0.11.8
tqdm==4.65.0
twine==3.8.0
typing_extensions==4.9.0
ujson==5.9.0
urllib3==1.26.15
uvicorn==0.25.0
uvloop==0.19.0
watchfiles==0.21.0
webencodings==0.5.1
websockets==10.4
zipp==3.15.0
